+++ src/MyLeague.cs
@@ -9,17 +9,17 @@ namespace BasketballSim
 
         private List<Game> leagueGames = null;
         private int leagueYear = 0;
-        private Stats leagueStats = null;
         private List<Team> leagueTeams = null;
         private int numDays;
         private int currentDay = 0;
         private Dictionary<Game, int> gameSchedule = null;
+        private Dictionary<Player, int> leaguePlayers = null;
 
         MyLeague(){
             leagueGames = new List<Game>();
             leagueTeams = new List<Team>();
-            leagueStats = new Stats();
             gameSchedule = new Dictionary<Game, int>();
+            leaguePlayers = new Dictionary<Player, int>();
         }
 
         private static readonly object padlock = new object();  
@@ -125,6 +125,32 @@ namespace BasketballSim
                 Console.WriteLine(t.getName()+": "+t.getWins().ToString()+"-"+t.getLosses().ToString());
             }
         }
+
+        public void addPoints(Dictionary<Player, PlayerStats> boxScore){
+            foreach (KeyValuePair<Player, PlayerStats> kvp in boxScore){
+                if(leaguePlayers.ContainsKey(kvp.Key)){
+                    leaguePlayers[kvp.Key] += kvp.Value.points;
+                    //leaguePlayers[kvp.Key] += kvp.Value;
+                }else{
+                    leaguePlayers.Add(kvp.Key,kvp.Value.points);
+                }
+            }
+        }
+
+        public void showStats(int totalGamesPlayed){
+            double ppg = 0;
+            foreach (KeyValuePair<Player, int> kvp in leaguePlayers){
+                ppg = kvp.Value / totalGamesPlayed;
+                Console.WriteLine(kvp.Key.getName() + " Anotó un total de: " + kvp.Value.ToString() + ". Promediando: " + ppg.ToString() + " por partido");
+            }
+                
+        }
+
+        public void getTopScorer(){
+            Player topScorer = leaguePlayers.Aggregate((l, r) => l.Value > r.Value ? l : r).Key;
+            Console.WriteLine("{0} is the League's Top Scorer with a total of {1} points and XX PPG", topScorer.getName(),leaguePlayers[topScorer]);
+        }
+
 
 
     }
