--- src/Stats.cs
@@ -1,19 +1,0 @@
-using System;
-using System.Collections.Generic;
-
-namespace BasketballSim
-{
-    public class Stats
-    {
-        private Dictionary<Player, int> playerPoints = null;
-        private Dictionary<Team, int> teamWins = null;
-        private Dictionary<Team, int> teamLosses = null;
-        public Stats(){
-            //<>
-        }
-        public string ReturnMessage()
-        {
-            return "code snippet para crear una clase nueva";
-        }
-    }
-}
